requirements.txt
# ── Build tools ───────────────────────────────────────────────────────────────
setuptools>=68.0.0
wheel>=0.41.0

# ── Core Orchestration ────────────────────────────────────────────────────────
# crewai 1.0+ removed the pkg_resources import that crashes on Windows
# If you get "No module named 'pkg_resources'", run: python patch_crewai.py
crewai>=1.0.0,<2.0.0
crewai-tools>=0.17.0,<0.20.0
langchain>=0.3.0,<0.4.0
langchain-groq>=0.2.0,<0.3.0

# ── LLM & AI ──────────────────────────────────────────────────────────────────
groq>=0.11.0,<0.12.0
faster-whisper>=1.0.0,<2.0.0
onnxruntime>=1.17.0
transformers>=4.44.0,<5.0.0
torch>=2.1.2,<2.2.0

# ── Text-to-Speech ────────────────────────────────────────────────────────────
kokoro>=0.9.0,<1.0.0
soundfile>=0.12.0
gtts>=2.5.0

# ── Video / Audio Assembly ────────────────────────────────────────────────────
moviepy>=1.0.3,<2.0.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0,<11.0.0
numpy>=1.26.4,<2.0.0
scipy>=1.14.0

# ── Media & Research ──────────────────────────────────────────────────────────
tavily-python>=0.5.0
requests>=2.32.0

# ── YouTube Upload ────────────────────────────────────────────────────────────
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
grpcio>=1.60.0

# ── Scheduling & Utilities ────────────────────────────────────────────────────
schedule>=1.2.0
python-dotenv>=1.0.0
pydantic>=2.4.2,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
loguru>=0.7.0
click>=8.1.0
rich>=13.0.0
httpx>=0.23.0,<0.28.0

# ── State Storage ─────────────────────────────────────────────────────────────
tinydb>=4.8.0
